refactor model tests
