--- app/tests/models/utils.py
@@ -1,15 +1,0 @@
-import os
-
-import pytest
-
-from app import create_app, db
-
-@pytest.fixture
-def context():
-    app = create_app()
-    with app.app_context() as context:
-        db.session.remove()
-        db.drop_all()
-        db.create_all()
-        yield context
-
